# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chatbay Analyzer API",
    "version": "5.8.0",
    "description": "Use ONLY when asked to analyze product photos or generate eBay CSVs. Do not call for general questions."
  },
  "servers": [{ "url": "https://chatbay-analyzer.onrender.com", "description": "Render prod" }],
  "paths": {
    "/health": { "get": { "summary": "Check API status", "operationId": "get_health_status", "responses": { "200": { "description": "OK" } } } },
    "/gallery": { "get": { "summary": "Fetch grouped image sets from WordPress", "operationId": "get_gallery_groups", "responses": { "200": { "description": "Groups" } } } },
    "/preview_csv": { "get": { "summary": "Preview first two analyzed listings (JSON only)", "operationId": "preview_csv", "parameters": [
      { "name": "photos_per_item", "in": "query", "required": true, "schema": { "type": "integer", "minimum": 1, "maximum": 12 } },
      { "name": "condition", "in": "query", "required": true, "schema": { "type": "string", "enum": ["new","preowned","parts"] } }
    ], "responses": { "200": { "description": "Preview JSON" } } } },
    "/export_csv": { "get": { "summary": "Generate and download CSV (browser attachment). For chat, prefer /export_csv_text or ?inline=1.", "operationId": "export_csv", "parameters": [
      { "name": "photos_per_item", "in": "query", "required": true, "schema": { "type": "integer", "minimum": 1, "maximum": 12 } },
      { "name": "condition", "in": "query", "required": true, "schema": { "type": "string", "enum": ["new","preowned","parts"] } },
      { "name": "inline", "in": "query", "required": false, "schema": { "type": "string", "enum": ["0","1"] } }
    ], "responses": { "200": { "description": "CSV file or inline JSON" } } } },
    "/export_csv_text": { "get": { "summary": "Generate CSV as JSON payload for chat surfaces", "operationId": "export_csv_text", "parameters": [
      { "name": "photos_per_item", "in": "query", "required": true, "schema": { "type": "integer", "minimum": 1, "maximum": 12 } },
      { "name": "condition", "in": "query", "required": true, "schema": { "type": "string", "enum": ["new","preowned","parts"] } }
    ], "responses": { "200": { "description": "Returns { filename, csv_text }" } } } }
  }
}